Parse should not raise an ArgumentError
IO.iodata_to_binary used in Parse will raise an error if the input is
not iodata, such as if we try to call Poison.decode(:hello). If an
error is raised, it will now be rescued and an {:error, :argument}
tuple will be returned instead. Parse! will raise an ArgumentError as
expected.