Added implementation of Encoder for Range
The previous implementation was encoding a range as a list of possible
values. This is not a viable approach, as it can possibly lead to huge
messages.
The new implementation treats a range as it is: a map, with a :first and
a :last fields. Thus, the resulting bitstring of Poison.encode(a..b) is
the same as Poison.encode(%{first: a, last: b}).